rename general test classes
